print

+++ tmpdir/external_func.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int external_func(int a, int b, int c, int d, int e, int f) {
+long external_func(int a, int b, int c, int d, int e, int f) {
   printf("hello! %d\n", a + 2 * b + 3 * c + 4 * d + 5 * e + 6 * f);
   return 5;
 }
